Extracts hooks to separate file, updates registration network call
