+++ src/components/auth/LoginForm.jsx
@@ -1,9 +1,8 @@
-import React, { useState, useContext } from 'react'
+import React from 'react'
 import { withRouter } from 'react-router-dom'
-import { AppContext } from '../../AppContext'
-import axios from '../../helpers/axiosService'
-import { setToken } from '../../helpers/localStorage'
+
 import AuthFormContainer from './AuthFormContainer'
+import { useLoginForm } from './authHooks'
 
 const LoginForm = (props) => {
   const [submitLogin, formData, onInputChangeHandler] = useLoginForm(props)
@@ -49,30 +48,3 @@ const LoginForm = (props) => {
 }
 
 export default withRouter(LoginForm)
-
-const useLoginForm = (props) => {
-  const [state, setState] = useContext(AppContext)
-  const [formData, setFormData] = useState({})
-  const onInputChangeHandler = (e) => {
-    setFormData({ ...formData, [e.target.name]: e.target.value })
-  }
-  const submitLogin = async (e) => {
-    e.preventDefault()
-    const payload = { user: formData }
-    try {
-      const response = await axios.login(payload)
-      setToken(response.data.user.token)
-      const currentUser = {
-        isSignedIn: true,
-        username: response.data.user.username,
-      }
-      setState((state) => ({ ...state, currentUser: currentUser }))
-      props.history.push({
-        pathname: '/',
-      })
-    } catch (error) {
-      console.error(error)
-    }
-  }
-  return [submitLogin, formData, onInputChangeHandler]
-}
