+++ src/components/auth/RegistrationForm.jsx
@@ -1,8 +1,8 @@
-import React, { useState, useContext } from 'react'
+import React from 'react'
 import { withRouter } from 'react-router-dom'
-import { AppContext } from '../../AppContext'
+
 import AuthFormContainer from './AuthFormContainer'
-const axios = require('axios')
+import { useRegistrationForm } from './authHooks'
 
 const RegistrationForm = (props) => {
   const [
@@ -61,42 +61,3 @@ const RegistrationForm = (props) => {
 }
 
 export default withRouter(RegistrationForm)
-
-const useRegistrationForm = (props) => {
-  const [state, setState] = useContext(AppContext)
-  const [formData, setFormData] = useState({})
-
-  const onInputChangeHandler = (e) => {
-    setFormData({ ...formData, [e.target.name]: e.target.value })
-  }
-  const submitRegistration = async (e) => {
-    e.preventDefault()
-    const formData = new FormData(e.target)
-    const payload = { user: Object.fromEntries(formData) }
-
-    try {
-      const response = await axios.post(
-        process.env.NODE_ENV === 'production'
-          ? 'https://conduit.productionready.io/api/users'
-          : 'http://localhost:3000/api/users',
-        JSON.stringify(payload),
-        {
-          headers: {
-            'Content-Type': 'application/json',
-          },
-        }
-      )
-      const currentUser = {
-        isSignedIn: true,
-        username: response.data.user.username,
-      }
-      setState((state) => ({ ...state, currentUser: currentUser }))
-      props.history.push({
-        pathname: '/',
-      })
-    } catch (error) {
-      console.error(error)
-    }
-  }
-  return [submitRegistration, onInputChangeHandler, formData]
-}
